Renamed all services' providers.
There was an unique provider for all services: "DecisionDeck".  The
situation has changed, now the services' providers are named after the
project they result from.  We here follow the move that is in progress
in the XMCDA web-services.

Cnsequently and besides their names, the content of every description
has changed, in that the webservices' names have been changed as well:
instead of being simply named after the program's name, they are now
called <name>-<provider> (for example: generateXMCDAReport-RXMCDA).
Please refer to the descriptions being renamed in this commit for an
overview of the renaming process.

A new delegate has been added, ProgramDelegate, that is responsible
for adapting the existing workflows so that their programs are
transparently migrated from the old name to the new one.

For the moment being the user is kept informed of the migration; this
is for verification purposes.  It is very likely that these warnings
won't be kept in the next release, as every program is concerned by
this renaming.