diviz@ECP now declares two platforms: ECP + uni.lu
The server installed at ECP now declares to execution platform, one
located at ECP and the other one at the University of Luxembourg.

Descriptions have been modified to reference a variable instead of
directly a script name; this way, platforms are able to select the URL
for the services they declare (they use the variable $callXMCDAService
for this).

NB: there is currently some network configuration problem on ddeck@ecp
so we sue a different configuration for the components deployed at
ddeck (pointing to localhost) and for the components deployed
elsewhere (typically the standalone version, in this case the
conf. points to ddeck's FQDN).