Added: a button to open a result in a HTML browser (to print it e.g.).
There is no direct support in cobra for printing, at least being able
to open results in an external browser allows the user to print it.

The generated HTML file has images pointing to the same directory
where it is stored (this was strangly not the case before, so these
HTML files were not displaying any images when opened in an external
browser).

A mechanism is provided to automatically adapt the existing results
(and their generated html file, specifically).

Last, the CSS has been modified so that images are displayed within
the browser frame (max-width: 100%).