diviz could not use any web-services at all in some cases.
In certain cases related to the network configuration of the user's computer,
diviz could not be used at all: all web-services were marked as unknown,
displayed in red.

We particularly thank Sascha Graßhoff for reporting the problem and for
working with us to solve it.

(bug present in Praxis v1.4, fixed in Praxis v1.4.1)