Revert "RMI: preferring direct connection to http tunnelling, if possible"
This reverts commit de5f0ff3c800a1b639fd60a3547325cc3deb8a75.

The commit re-introduced a problem that have been previously solved,
where some cleints could not connect at all if rmi.useHTTP is false.
The problem is that under some (yet unclear) circumstances, the
attempt to establish a RMI connection through a standard socket does
not timeout, so the HTTP fallback is never tried and the client's
attempt to connect just hangs.